requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Vector Database
chromadb==0.4.22

# Embeddings & LLM
openai==1.7.2
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
